+++ src/joints/box2d_joint.cpp
@@ -14,6 +14,12 @@
 
 void Box2DJoint::clear() {
 	configured = false;
+	if (common.body_a) {
+		common.body_a->remove_joint(this);
+	}
+	if (common.body_b) {
+		common.body_b->remove_joint(this);
+	}
 	common.body_a = nullptr;
 	common.body_b = nullptr;
 	if (space) {
@@ -111,24 +117,24 @@ real_t Box2DJoint::get_pin_softness() {
 }
 
 void Box2DJoint::set_damped_spring_rest_length(real_t p_damped_spring_rest_length) {
-	if (damped_spring.rest_length == godot_to_box2d(godot_to_box2d(p_damped_spring_rest_length))) {
+	if (damped_spring.rest_length == godot_to_box2d(p_damped_spring_rest_length)) {
 		return;
 	}
-	damped_spring.rest_length = godot_to_box2d(godot_to_box2d(p_damped_spring_rest_length));
+	damped_spring.rest_length = godot_to_box2d(p_damped_spring_rest_length);
 	if (joint && (b2DistanceJoint *)joint) {
 		b2DistanceJoint *distance_joint = (b2DistanceJoint *)joint;
 		distance_joint->SetLength(damped_spring.rest_length);
 	}
 }
 real_t Box2DJoint::get_damped_spring_rest_length() {
-	return box2d_to_godot(box2d_to_godot(damped_spring.rest_length));
+	return box2d_to_godot(damped_spring.rest_length);
 }
 
 void Box2DJoint::set_damped_spring_stiffness(real_t p_damped_spring_stiffness) {
-	if (damped_spring.stiffness == godot_to_box2d(godot_to_box2d(p_damped_spring_stiffness))) {
+	if (damped_spring.stiffness == godot_to_box2d(p_damped_spring_stiffness)) {
 		return;
 	}
-	damped_spring.stiffness = godot_to_box2d(godot_to_box2d(p_damped_spring_stiffness));
+	damped_spring.stiffness = godot_to_box2d(p_damped_spring_stiffness);
 	if (joint && (b2DistanceJoint *)joint && common.body_a && common.body_b) {
 		b2DistanceJoint *distance_joint = (b2DistanceJoint *)joint;
 		float stiffness = 0;
@@ -138,7 +144,7 @@ void Box2DJoint::set_damped_spring_stiffness(real_t p_damped_spring_stiffness) {
 	}
 }
 real_t Box2DJoint::get_damped_spring_stiffness() {
-	return box2d_to_godot(box2d_to_godot(damped_spring.stiffness));
+	return box2d_to_godot(damped_spring.stiffness);
 }
 
 void Box2DJoint::set_damped_spring_damping(real_t p_damped_spring_damping) {
@@ -178,7 +184,6 @@ b2JointDef *Box2DJoint::get_b2JointDef() {
 			distance_joint_def->Initialize(common.body_a->get_b2Body(), common.body_b->get_b2Body(), common.anchor_a, common.anchor_b);
 
 			distance_joint_def->length = damped_spring.rest_length;
-			distance_joint_def->minLength = 0;
 		} break;
 		case PhysicsServer2D::JointType::JOINT_TYPE_GROOVE: {
 			b2PrismaticJointDef *prismatic_joint_def = (b2PrismaticJointDef *)joint_def;
@@ -188,7 +193,7 @@ b2JointDef *Box2DJoint::get_b2JointDef() {
 			prismatic_joint_def->enableLimit = true;
 		} break;
 		default: {
-			ERR_PRINT("UNSUPPORTED");
+			ERR_PRINT("Unkown joint type");
 		}
 	}
 	return joint_def;
@@ -209,5 +214,6 @@ Box2DJoint::Box2DJoint() {
 }
 
 Box2DJoint::~Box2DJoint() {
+	clear();
 	memdelete(joint_def);
 }