+++ src/shapes/box2d_shape_rectangle.cpp
@@ -15,21 +15,26 @@ void Box2DShapeRectangle::set_data(const Variant &p_data) {
 		half_extents.y = GODOT_LINEAR_SLOP;
 	}
 	configured = true;
+	// update all existing shapes
+	configure_all_b2Shapes();
 }
 
 Variant Box2DShapeRectangle::get_data() const {
 	return half_extents;
 }
-
-b2Shape *Box2DShapeRectangle::get_transformed_b2Shape(int p_index, Transform2D &p_transform, bool one_way, bool is_static) {
-	ERR_FAIL_INDEX_V(p_index, 1, nullptr);
+b2Shape *Box2DShapeRectangle::get_transformed_b2Shape(ShapeInfo shape_info, Box2DCollisionObject *body) {
+	ERR_FAIL_INDEX_V(shape_info.index, 1, nullptr);
 	b2PolygonShape *shape = memnew(b2PolygonShape);
+	created_shapes.append(shape);
+	if (body) {
+		shape_body_map[shape] = body;
+	}
 	b2Vec2 box2d_half_extents = godot_to_box2d(half_extents);
 	b2Vec2 *box2d_points = new b2Vec2[4];
-	godot_to_box2d(p_transform.xform(Vector2(-half_extents.x, -half_extents.y)), box2d_points[0]);
-	godot_to_box2d(p_transform.xform(Vector2(-half_extents.x, half_extents.y)), box2d_points[1]);
-	godot_to_box2d(p_transform.xform(Vector2(half_extents.x, half_extents.y)), box2d_points[2]);
-	godot_to_box2d(p_transform.xform(Vector2(half_extents.x, -half_extents.y)), box2d_points[3]);
+	godot_to_box2d(shape_info.transform.xform(Vector2(-half_extents.x, -half_extents.y)), box2d_points[0]);
+	godot_to_box2d(shape_info.transform.xform(Vector2(-half_extents.x, half_extents.y)), box2d_points[1]);
+	godot_to_box2d(shape_info.transform.xform(Vector2(half_extents.x, half_extents.y)), box2d_points[2]);
+	godot_to_box2d(shape_info.transform.xform(Vector2(half_extents.x, -half_extents.y)), box2d_points[3]);
 	shape->Set(box2d_points, 4);
 	delete[] box2d_points;
 	return shape;